--- src/components/pagination/PaginationDefault.tsx
@@ -1,75 +1,0 @@
-import * as React from 'react';
-import {createContext, FC, HTMLAttributes, useMemo} from 'react';
-import {Item} from './Item';
-import {Ellipsis} from './Ellipsis';
-import {Next} from './Next';
-import {Prev} from './Prev';
-import cx from 'classnames';
-import {component} from '../../services/helpers/classHelpers';
-import {Container} from './Container';
-
-import './cpagination.less';
-
-export type TOnChange = (page: number) => void;
-
-export interface IPaginationDefault extends Omit<HTMLAttributes<HTMLDivElement>, 'onChange'> {
-    theme?: 'dark' | 'light';
-    activePage: number;
-    total: number;
-    onChange?: TOnChange;
-}
-
-export interface IPaginationDefaultContext {
-    activePage: number;
-    total: number;
-    onChange: TOnChange;
-}
-
-const defaultChange = () => {
-    //
-};
-
-export const PaginationContext = createContext<IPaginationDefaultContext>({
-    activePage: 0,
-    total: 0,
-    onChange: defaultChange,
-});
-
-const BasePagination: FC<IPaginationDefault> = ({
-    activePage,
-    total,
-    children,
-    onChange,
-    theme = 'light',
-    className,
-    ...restProps
-}) => {
-    const paginationClassName = useMemo<string>(
-        () =>
-            cx(
-                component('pagination')({
-                    [theme]: true,
-                }),
-                className,
-            ),
-        [className, theme],
-    );
-
-    return (
-        <div {...restProps} className={paginationClassName}>
-            <PaginationContext.Provider value={{activePage, total, onChange: onChange || defaultChange}}>
-                {children}
-            </PaginationContext.Provider>
-        </div>
-    );
-};
-
-const paginationComponents = {
-    Item,
-    Prev,
-    Next,
-    Container,
-    Ellipsis,
-};
-
-export const PaginationDefault = Object.assign(BasePagination, paginationComponents);
