+++ src/components/pagination/Prev.tsx
@@ -3,7 +3,7 @@ import {FC, HTMLAttributes, useContext, useMemo} from 'react';
 import cx from 'classnames';
 import {component} from '../../services/helpers/classHelpers';
 import Left from '../../assets/svg/48/shape-left-48.svg';
-import {PaginationContext} from './PaginationDefault';
+import {PaginationContext} from './Pagination';
 
 export const Prev: FC<HTMLAttributes<HTMLButtonElement>> = ({className, ...restProps}) => {
     const {activePage, onChange} = useContext(PaginationContext);
